SCRUM-106 fix: ClientPage로 분리
