+++ frontend/src/app/question/list/page.tsx
@@ -16,7 +16,6 @@ function convertSnakeToCamel<T>(obj: T): T {
   return obj;
 }
 
-type QuestionDto = components["schemas"]["QuestionDto"];
 type PageDtoQuestionDto = components["schemas"]["PageDtoQuestionDto"];
 
 export default async function Page() {
@@ -30,28 +29,5 @@ export default async function Page() {
   const body: PageDtoQuestionDto = convertSnakeToCamel(data);
   console.log("API 응답:", body); // 응답 확인
 
-  return (
-    <div>
-      <div>currentPageNumber: {body.currentPageNumber}</div>
-      <div>pageSize: {body.pageSize}</div>
-      <div>totalPages: {body.totalPages}</div>
-      <div>totalItems: {body.totalItems}</div>
-      
-      <hr /><br />
-      <ul>
-        {body.items?.map((item: QuestionDto) => (
-          <li key={item.id}>
-            <div>id: {item.id}</div>
-            <div>title: {item.title}</div>
-            <div>content: {item.content}</div>
-            <div>authorName: {item.name}</div>
-            <div>createdAt: {item.createdAt}</div>
-            <div>modifiedAt: {item.modifiedAt}</div><br />
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-
-  // return <ClientPage />;
+  return <ClientPage body={body} />;
 }
